moving createPerformanceMeasure to tsx file to appease the dependency walker
