+++ editor/src/components/editor/store/store-hook-performance-logging.ts
@@ -3,8 +3,6 @@ import { PERFORMANCE_MARKS_ALLOWED } from '../../../common/env-vars'
 import { isFeatureEnabled } from '../../../utils/feature-switches'
 import { StateSelector } from './store-hook'
 import { StoreKey, Substates } from './store-hook-substore-types'
-import { simpleStringifyActions } from './dispatch'
-import { EditorAction } from '../action-types'
 
 export function resetSelectorTimings(): void {
   SelectorTimings.current = {}
@@ -181,63 +179,4 @@ export function logSelectorTimings(phase: string): void {
   console.log(`Pre-Selectors during ${phase}`)
   // eslint-disable-next-line no-console
   console.table(SubstoreTimings.current, ['accumulatedTime', 'calledNumberOfTimes'])
-}
-
-export function createPerformanceMeasure() {
-  const MeasureSelectorsEnabled = isFeatureEnabled('Debug – Measure Selectors')
-  const PerformanceMarks =
-    (isFeatureEnabled('Debug – Performance Marks (Slow)') ||
-      isFeatureEnabled('Debug – Performance Marks (Fast)')) &&
-    PERFORMANCE_MARKS_ALLOWED
-
-  let stringifiedActions = ''
-
-  let tasks: {
-    [name: string]: { timingInfo: string; startMarkName: string; endMarkName: string }
-  } = {}
-  return {
-    logActions(actions: ReadonlyArray<EditorAction>): void {
-      stringifiedActions = simpleStringifyActions(actions)
-    },
-    taskTime<T>(name: string, task: () => T): T {
-      const startMarkName = `${name}-start`
-      const endMarkName = `${name}-end`
-
-      const before = MeasureSelectorsEnabled ? performance.now() : 0
-      if (PerformanceMarks) {
-        performance.mark(startMarkName)
-      }
-
-      // run the actual task
-      const result = task()
-
-      const after = MeasureSelectorsEnabled ? performance.now() : 0
-      if (PerformanceMarks) {
-        performance.mark(endMarkName)
-      }
-
-      tasks[name] = {
-        timingInfo: `${after - before}ms`,
-        startMarkName: startMarkName,
-        endMarkName: endMarkName,
-      }
-      return result
-    },
-    printMeasurements() {
-      if (MeasureSelectorsEnabled) {
-        console.info('------------------')
-        console.info(`dispatched actions: ${stringifiedActions}`)
-        console.info(
-          Object.entries(tasks)
-            .map(([task, { timingInfo }]) => `${task}: ${timingInfo}`)
-            .join('\n'),
-        )
-      }
-      if (PerformanceMarks) {
-        Object.entries(tasks).forEach(([task, { startMarkName, endMarkName }]) => {
-          performance.measure(task, startMarkName, endMarkName)
-        })
-      }
-    },
-  }
 }
